Move toUTCDate to date.ts
